refactor:: 스케줄 관련 리팩토링 - schedule service에서 불필요한 try-catch 구문 삭제 - 시간충돌 검증 로직 중복되어있어서 utils로 빼놓고 불러와서 사용함으로써 유지보수성과 재사용성 확보
