+++ src/schedule/dto/delete-schedule-response.dto.ts
@@ -3,4 +3,8 @@ import { ApiProperty } from '@nestjs/swagger';
 export class DeleteScheduleResponseDto {
   @ApiProperty({ description: '삭제 여부' })
   deleted: boolean;
+
+  constructor(deleted: boolean) {
+    this.deleted = deleted;
+  }
 }
