+++ src/schedule/dto/update-schedule-response.dto.ts
@@ -1,5 +1,4 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsNotEmpty, IsNumber, IsString } from 'class-validator';
 
 export class UpdateScheduleResponseDto {
   @ApiProperty({ description: 'ID' })
